+++ dashboard/app/page.tsx
@@ -1,7 +1,8 @@
 import database from "./data/wanhua-community-associations.json";
 import activities from "./data/wanhua-community-activities.json";
-import { WanhuaDashboard, type ActivityStatistics, type Database } from "./components/WanhuaDashboard";
+import sdgCandidates from "./data/wanhua-community-sdg-candidates.json";
+import { WanhuaDashboard, type ActivityStatistics, type Database, type SDGCandidateStatistics } from "./components/WanhuaDashboard";
 
 export default function Home() {
-  return <WanhuaDashboard database={database as Database} activities={activities as ActivityStatistics} googleMapsApiKey={process.env.GOOGLE_MAPS_API_KEY} />;
+  return <WanhuaDashboard database={database as Database} activities={activities as ActivityStatistics} sdgCandidates={sdgCandidates as SDGCandidateStatistics} googleMapsApiKey={process.env.GOOGLE_MAPS_API_KEY} />;
 }